+++ geolocator_platform_interface/lib/src/models/position.dart
@@ -6,17 +6,17 @@ class Position {
   /// Constructs an instance with the given values for testing. [Position]
   /// instances constructed this way won't actually reflect any real information
   /// from the platform, just whatever was passed in at construction time.
-  Position({
-    this.longitude,
-    this.latitude,
-    this.timestamp,
-    this.accuracy,
-    this.altitude,
-    this.heading,
+  const Position({
+    required this.longitude,
+    required this.latitude,
+    required this.timestamp,
+    required this.accuracy,
+    required this.altitude,
+    required this.heading,
+    required this.speed,
+    required this.speedAccuracy,
     this.floor,
-    this.speed,
-    this.speedAccuracy,
-    this.isMocked,
+    this.isMocked = false,
   });
 
   /// The latitude of this position in degrees normalized to the interval -90.0
@@ -28,7 +28,7 @@ class Position {
   final double longitude;
 
   /// The time at which this position was determined.
-  final DateTime timestamp;
+  final DateTime? timestamp;
 
   /// The altitude of the device in meters.
   ///
@@ -53,7 +53,7 @@ class Position {
   ///
   /// The floor property is only available on iOS and only when the information
   /// is available. In all other cases this value will be null.
-  final int floor;
+  final int? floor;
 
   /// The speed at which the devices is traveling in meters per second over
   /// ground.
@@ -75,18 +75,18 @@ class Position {
   final bool isMocked;
 
   @override
-  bool operator ==(dynamic o) {
-    var areEqual = o is Position &&
-        o.accuracy == accuracy &&
-        o.altitude == altitude &&
-        o.heading == heading &&
-        o.latitude == latitude &&
-        o.longitude == longitude &&
-        o.floor == o.floor &&
-        o.speed == speed &&
-        o.speedAccuracy == speedAccuracy &&
-        o.timestamp == timestamp &&
-        o.isMocked == isMocked;
+  bool operator ==(Object other) {
+    var areEqual = other is Position &&
+        other.accuracy == accuracy &&
+        other.altitude == altitude &&
+        other.heading == heading &&
+        other.latitude == latitude &&
+        other.longitude == longitude &&
+        other.floor == floor &&
+        other.speed == speed &&
+        other.speedAccuracy == speedAccuracy &&
+        other.timestamp == timestamp &&
+        other.isMocked == isMocked;
 
     return areEqual;
   }
@@ -111,10 +111,6 @@ class Position {
 
   /// Converts the supplied [Map] to an instance of the [Position] class.
   static Position fromMap(dynamic message) {
-    if (message == null) {
-      return null;
-    }
-
     final Map<dynamic, dynamic> positionMap = message;
 
     if (!positionMap.containsKey('latitude')) {