+++ geolocator_platform_interface/test/src/implementations/event_channel_mock.dart
@@ -1,19 +1,17 @@
 import 'dart:async';
 
-import 'package:flutter/cupertino.dart';
 import 'package:flutter/services.dart';
-import 'package:meta/meta.dart';
 
 class EventChannelMock {
   final MethodChannel _methodChannel;
-  final Stream stream;
   final log = <MethodCall>[];
 
-  StreamSubscription _streamSubscription;
+  Stream? stream;
+  StreamSubscription? _streamSubscription;
 
   EventChannelMock({
-    @required String channelName,
-    @required this.stream,
+    required String channelName,
+    required this.stream,
   }) : _methodChannel = MethodChannel(channelName) {
     _methodChannel.setMockMethodCallHandler(_handler);
   }
@@ -29,37 +27,31 @@ class EventChannelMock {
         _onCancel();
         break;
       default:
-        return null;
+        return Future.value(null);
     }
 
     return Future.value();
   }
 
   void _onListen() {
-    _streamSubscription = stream.handleError((error) {
-      ServicesBinding.instance.defaultBinaryMessenger.handlePlatformMessage(
-        _methodChannel.name,
-        _createErrorEnvelope(error),
-        (_) {},
-      );
-    }).listen((event) {
-      ServicesBinding.instance.defaultBinaryMessenger.handlePlatformMessage(
-        _methodChannel.name,
-        _createSuccessEnvelope(event),
-        (_) {},
-      );
-    });
+    _streamSubscription = stream!.handleError(_sendErrorEnvelope).listen(
+      _sendSuccessEnvelope,
+      onDone: () {
+        _sendEnvelope(null);
+      },
+    );
   }
 
   void _onCancel() {
     if (_streamSubscription != null) {
-      _streamSubscription.cancel();
+      _streamSubscription!.cancel();
+      stream = null;
     }
   }
 
-  ByteData _createErrorEnvelope(Exception error) {
+  void _sendErrorEnvelope(Exception error) {
     var code = "UNKNOWN_EXCEPTION";
-    String message;
+    String? message;
     dynamic details;
 
     if (error is PlatformException) {
@@ -68,11 +60,26 @@ class EventChannelMock {
       details = error.details;
     }
 
-    return const StandardMethodCodec()
+    final envelope = const StandardMethodCodec()
         .encodeErrorEnvelope(code: code, message: message, details: details);
+
+    _sendEnvelope(envelope);
   }
 
-  ByteData _createSuccessEnvelope(dynamic event) {
-    return const StandardMethodCodec().encodeSuccessEnvelope(event);
+  void _sendSuccessEnvelope(dynamic event) {
+    final envelope = const StandardMethodCodec().encodeSuccessEnvelope(event);
+    _sendEnvelope(envelope);
+  }
+
+  void _sendEnvelope(ByteData? envelope) {
+    if (ServicesBinding.instance == null) {
+      return;
+    }
+
+    ServicesBinding.instance!.defaultBinaryMessenger.handlePlatformMessage(
+      _methodChannel.name,
+      envelope,
+      (_) {},
+    );
   }
 }